Return course id
